+++ hybrid_core/hybrid_core_android/lib/src/plugin.dart
@@ -1,9 +1,11 @@
 import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
 
+import 'jni.dart';
 import 'os.dart';
 
-abstract class HybridCoreWindows {
+abstract class HybridCoreAndroidPlugin {
   static void registerWith() {
-    OSPlatform.instance = WindowsPlatform();
+    Jni.initDLApi();
+    OSPlatform.instance = AndroidPlatform();
   }
 }