+++ hybrid_core/hybrid_core_darwin/lib/hybrid_core_darwin.dart
@@ -1,2 +1,2 @@
-export 'src/hybrid_core_darwin.dart';
+export 'src/plugin.dart';
 export 'src/os.dart';