+++ hybrid_core/hybrid_core_darwin/lib/src/ffi.dart
@@ -2,12 +2,6 @@ import 'dart:ffi';
 
 import 'ffi.g.dart';
 
-export 'dart:ffi';
-
-export 'package:ffi/ffi.dart';
-
-export 'ffi.g.dart';
-
 const foundationPath =
     '/System/Library/Frameworks/Foundation.framework/Foundation';
 final foundationLib = DynamicLibrary.open(foundationPath);