+++ hybrid_core/hybrid_core_darwin/lib/src/plugin.dart
@@ -1,0 +1,12 @@
+import 'dart:io';
+
+import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
+
+import 'os.dart';
+
+abstract class HybridCoreDarwinPlugin {
+  static void registerWith() {
+    OSPlatform.instance =
+        Platform.isIOS ? iOSPlatformImpl() : macOSPlatformImpl();
+  }
+}