+++ hybrid_core/hybrid_core_platform_interface/lib/hybrid_core_platform_interface.dart
@@ -1,1 +1,3 @@
+export 'src/logging.dart';
 export 'src/os.dart';
+export 'src/ui.dart';