+++ hybrid_core/hybrid_core_windows/lib/src/ffi.dart
@@ -1,16 +1,11 @@
-// ignore_for_file: type=lint
+// ignore_for_file: non_constant_identifier_names
+
 import 'dart:ffi';
 
 import 'ffi.g.dart';
 
-export 'dart:ffi';
-
-export 'package:ffi/ffi.dart';
-
-export 'ffi.g.dart';
-
 final kernel32Lib = DynamicLibrary.open('Kernel32.dll');
-final kernel32 = HybridCore(kernel32Lib);
+final kernel32 = FFI(kernel32Lib);
 
 int LOBYTE(int w) => w & 0xff;
 int HIBYTE(int w) => (w >> 8) & 0xff;