+++ hybrid_core/hybrid_core_windows/lib/src/plugin.dart
@@ -1,11 +1,9 @@
-import 'dart:io';
-
 import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
 
 import 'os.dart';
 
-abstract class HybridCoreDarwin {
+abstract class HybridCoreWindowsPlugin {
   static void registerWith() {
-    OSPlatform.instance = Platform.isIOS ? iOSPlatform() : macOSPlatform();
+    OSPlatform.instance = WindowsPlatform();
   }
 }