+++ hybrid_core/hybrid_core_windows/CHANGELOG.md
@@ -1,7 +1,20 @@
-## 0.0.1
+## 0.3.2
 
-* Add support for Windows.
+* Move `WindowsHybridCorePlugin` to `HybridCoreWindowsPlugin`.
+
+## 0.3.1
+
+* Implement `hybrid_core_platform_interface` 0.3.1.
+
+## 0.3.0
+
+* Implement `hybrid_core_platform_interface` 0.3.1.
+* Move `HybridCoreWindows` to `WindowsHybridCorePlugin`.
 
 ## 0.2.0
 
 * Implement `OSPlatform` interface.
+
+## 0.0.1
+
+* Add support for Windows.